refactor: move db runner to src/infra and add domain AppError
