+++ src/domain/errors.ts
@@ -1,0 +1,8 @@
+export class AppError extends Error {
+  status: number;
+  constructor(status: number, message: string) {
+    super(message);
+    this.name = 'AppError';
+    this.status = status;
+  }
+}
