+++ src/infra/db.ts
@@ -1,18 +1,22 @@
-const fs = require('fs');
-const path = require('path');
-const Database = require('better-sqlite3');
+import fs from 'fs';
+import path from 'path';
+import Database from 'better-sqlite3';
 
-function openDatabase(filename) {
+export type Db = Database.Database;
+
+export function openDatabase(filename: string): Db {
   const db = new Database(filename);
   db.pragma('journal_mode = WAL');
   db.pragma('foreign_keys = ON');
   return db;
 }
 
-function runMigrations(db, dir = path.join(__dirname, '..', 'migrations')) {
+export function runMigrations(db: Db, dir = path.join(__dirname, '..', '..', 'migrations')): void {
   db.exec(`CREATE TABLE IF NOT EXISTS schema_migrations (
     name TEXT PRIMARY KEY, applied_at TEXT NOT NULL DEFAULT (datetime('now')));`);
-  const applied = new Set(db.prepare('SELECT name FROM schema_migrations').all().map(r => r.name));
+  const applied = new Set(
+    (db.prepare('SELECT name FROM schema_migrations').all() as { name: string }[]).map(r => r.name),
+  );
   const files = fs.readdirSync(dir).filter(f => f.endsWith('.sql')).sort();
   for (const file of files) {
     if (applied.has(file)) continue;
@@ -23,5 +27,3 @@ function runMigrations(db, dir = path.join(__dirname, '..', 'migrations')) {
     })();
   }
 }
-
-module.exports = { openDatabase, runMigrations };
