+++ test/crud.test.js
@@ -1,7 +1,7 @@
 const { test } = require('node:test');
 const assert = require('node:assert');
 const { makeTestDb } = require('./helpers');
-const { openDatabase, runMigrations } = require('../src/db');
+const { openDatabase, runMigrations } = require('../src/infra/db');
 const os = require('os');
 const path = require('path');
 const fs = require('fs');
